WIP TwoWay time binding Still not working angle calculation :(

+++ app/src/main/java/pl/brokenpipe/timeboxing/screens/timer/newtimer/interfaces/TimerViewModel.kt
@@ -9,5 +9,5 @@ import pl.brokenpipe.timeboxing.screens.timer.newtimer.TimerViewState
 interface TimerViewModel : ViewModel<TimerView, TimerViewState> {
     fun startTimer()
     fun pauseTimer()
-    fun setTimeByAngle(angle: Float)
+    fun setTimeByAngle(angle: Double)
 }
